Add energy
