+++ src/root/bonfire/bonfire.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { Resource } from '../interface';
-import { Observable } from 'rxjs';
 import { ResourceService } from '../resources/resource.service';
 
 @Component({
@@ -13,8 +11,6 @@ export class BonfireComponent implements OnInit {
   constructor(
     private resourcesService: ResourceService
   ) {}
-
-  private trashArray$: Observable<Map<string, Resource[]>>;
 
   ngOnInit(): void {
   }
