+++ src/root/resources/interface.ts
@@ -1,0 +1,24 @@
+export enum TYPES {
+  TRASH = 'Trash',
+  REFINED = 'Refined',
+  BASIC = ''
+}
+
+export enum TRASH {
+  RUBBISH = 0,
+  KINDLING = 1,
+  SCRAP = 2,
+  PLASTIC = 3,
+  ELECTRONIC = 4,
+  MEDICAL = 5
+}
+
+export enum REFINED {
+  PCHUNK = 0,
+}
+
+export enum Tasks {
+  NOTHING = 0,
+  REST = 1,
+  TEND_FIRE = 2
+}
