+++ src/root/resources/resources.component.scss
@@ -27,5 +27,8 @@
     &.metal {
       color: #9A9B94
     }
+    &.energy {
+      color: #ffc107;
+    }
   }
 }
